@@ -2,22 +2,22 @@
   "name": "student-skills-database-frontened",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
     "@types/react": "^17.0.39",
     "@types/react-dom": "^17.0.11",
     "emailjs-com": "^2.6.4",
-    "node-sass": "^4.14.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-redux": "^7.2.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.2",
-    "redux": "^4.0.5",
+    "node-sass": "^7.0.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.6",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^5.0.0",
+    "redux": "^4.1.2",
     "redux-devtools-extension": "^2.13.9",
     "smooth-scroll": "^16.1.3",
-    "typescript": "^3.7.5"
+    "typescript": "^4.5.5"
   },
   "scripts": {
     "start": "react-scripts start",
